+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0-nullsafety.1 - Jan 23, 2021
+
+*   **Breaking**: changed signature of `LoaderMessageHandler` to `void Function(BuildContext, LoaderMessage, LoaderBloc)`.
+    This allows using `BuildContext` to access ancestor widget, eg. `ScaffoldMessager.of(context)`, `Navigator.of(context)`, ...
+
 ## 1.1.0-nullsafety.0 - Jan 18, 2021
 *   Migrate this package to null safety.
 *   Sdk constraints: `>=2.12.0-0 <3.0.0` based on beta release guidelines.